% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malf_demographics.R
\docType{data}
\name{malf_demographics}
\alias{malf_demographics}
\title{Demographics for Subjects in MALF Data}
\format{A list with 7 elements, which are:
\describe{
\item{id}{identifier corresponding to the files }
\item{group}{training/test designation}
\item{Subject}{Subject name from original OASIS data set }
\item{Gender}{sex of subject}
\item{Age}{age of subject}
\item{Notes}{any notes on which scan was taken}
}}
\usage{
malf_demographics
}
\description{
A \code{data.frame} containing the demographic information for
the subjects in the training and test sets
}
\keyword{datasets}
