% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malf_labels.R
\docType{data}
\name{malf_labels}
\alias{malf_labels}
\title{Label Names for Labeled Images}
\format{A list with 3 elements, which are:
\describe{
\item{index}{integer relating to the labeled image}
\item{name}{structure name of the brain}
\item{ignore}{indicator if the label was "ignored" as indicated by the
challenge coordinators.  These may not be always labeled by each reader.}
}}
\usage{
malf_labels
}
\description{
A \code{data.frame} containing the label indices and
their corresponding names for identification of structures.
}
\keyword{datasets}
