% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_images.R
\name{mass_images}
\alias{mass_images}
\alias{mass_template_images}
\alias{mass_template_masks}
\title{Get Image Filenames for Skull Stripping}
\usage{
mass_images(n_templates = 35)

mass_template_images(...)

mass_template_masks(...)
}
\arguments{
\item{n_templates}{Number of template filenames to return}

\item{...}{arguments to pass \code{\link{mass_images}}}
}
\value{
List of filenames for images and masks
}
\description{
Return a list of filenames for the images and
brain masks
}
